+++ front-end/src/App.scss
@@ -1042,4 +1042,38 @@ footer {
 }
 /* generated by https://loading.io/ */
 
+.scrollBtn {
+	position: fixed;
+	bottom: 50px;
+	right: 20px;
+	padding: 10px;
+	outline: none;
+	background-color: $blue;
+	color: $black;
+	border: none;
+	cursor: pointer;
+	border-radius: 5px;
+	z-index: 100;
+	&:hover {
+		background-color: $blue-hover;
+		color: $black;
+	}
+}
+
+.scrollBtnMobile {
+	position: fixed;
+	bottom: 50px;
+	right: 13px;
+	// padding: 10px;
+	outline: none;
+	color: $blue;
+	border: none;
+	cursor: pointer;
+	border-radius: 5px;
+	font-size: 30px;
+	&:hover {
+		color: $blue-hover;
+	}
+}
+
 @import "media.scss";
